requirements.txt
Flask==3.0.0
flask-session==0.5.0
werkzeug==3.0.0
flask-cors==4.0.0
reportlab==4.0.7
openpyxl==3.1.5
psycopg2-binary
